delete new and old implementation of Cache
