+++ server/src/main/java/org/apache/iotdb/db/mpp/transformation/datastructure/row/ElasticSerializableRowRecordList.java
@@ -21,7 +21,7 @@ package org.apache.iotdb.db.mpp.transformation.datastructure.row;
 
 import org.apache.iotdb.db.exception.query.QueryProcessException;
 import org.apache.iotdb.db.mpp.transformation.dag.util.InputRowUtils;
-import org.apache.iotdb.db.mpp.transformation.datastructure.NewCache;
+import org.apache.iotdb.db.mpp.transformation.datastructure.Cache;
 import org.apache.iotdb.db.mpp.transformation.datastructure.SerializableList;
 import org.apache.iotdb.tsfile.file.metadata.enums.TSDataType;
 import org.apache.iotdb.tsfile.utils.Binary;
@@ -280,7 +280,7 @@ public class ElasticSerializableRowRecordList {
     this.evictionUpperBound = evictionUpperBound;
   }
 
-  private class LRUCache extends NewCache {
+  private class LRUCache extends Cache {
 
     LRUCache(int capacity) {
       super(capacity);
